Move CurveHandler base classes to Processors module
